public/talks/dataai2024/requirements.txt
fastapi
orjson
perspective-python==2.10.0
pyarrow
pyspark
superstore
uvicorn
websocket-client
ipywidgets
jupyterlab